New feature; User definable Clipping Planes.
Press ALT+B in 3d window, draw a rect, and it becomes a clipping
volume of 4 planes. You then can rotate the view anyway you like.
Works for each 3d window individually.

Disable it with another ALT+B press.

Commit is huge because it had to change all selection code as well.
The user-clipping planes are in 'eye space', the other clipping
happens in projected 'viewport space'.

Nice to notice is that the 'x=3200' convention (to denote a coordinate
is clipped) now is a define. Define value is still a number though... but
we now can get up to screens of 12000 pixels without issues!

Known issue; here it refuses to draw the 'object centers' or Lamp icons
within the clipping region. Can't find any reason for it... however, we
might move to non-pixmaps for it anyway.

Testing might reveil numerous issues, will be standby for it.

Curious? Check this http://www.blender.org/bf/rt4.jpg

+++ source/blender/src/editlattice.c
@@ -214,7 +214,7 @@ void deselectall_Latt(void)
 static void findnearestLattvert__doClosest(void *userData, BPoint *bp, int x, int y)
 {
 	struct { BPoint *bp; short dist, select, mval[2]; } *data = userData;
-	short temp = abs(data->mval[0]-x) + abs(data->mval[1]-y);
+	float temp = abs(data->mval[0]-x) + abs(data->mval[1]-y);
 	
 	if ((bp->f1&1)==data->select) temp += 5;
 	if (temp<data->dist) {
